Handle meta elements outside of head
