--- tests/wpt/meta/referrer-policy/generic/meta-referrer-outofhead-fetch.http.html.ini
@@ -1,3 +1,0 @@
-[meta-referrer-outofhead-fetch.http.html]
-  [Referer header was not send]
-    expected: FAIL
